refactor: amount,balanceの型をBigIntに統一

+++ app/adapter/src/main/scala/myapp/adapter/account/TransactionDto.scala
@@ -3,7 +3,7 @@ package myapp.adapter.account
 final case class TransactionDto(
     transactionId: String,
     transactionType: String,
-    amount: Int,
-    balance: Int,
+    amount: Long,
+    balance: Long,
     transactedAt: Long,
 )
