+++ app/application/src/main/scala/myapp/application/projection/transaction/Transaction.scala
@@ -8,6 +8,6 @@ final case class Transaction(
     eventType: TransactionEventType,
     accountNo: AccountNo,
     amount: BigInt,
-    balance: Int,
+    balance: BigInt,
     transactedAt: Long,
 )
